+++ src/tracks/interfaces/track.interface.ts
@@ -1,0 +1,9 @@
+interface Track {
+  id: string; // uuid v4
+  name: string;
+  artistId: string | null; // refers to Artist
+  albumId: string | null; // refers to Album
+  duration: number; // integer number
+}
+
+export { Track };
