+++ src/tracks/tracks.controller.ts
@@ -1,6 +1,7 @@
-import { Controller, Param, Res } from '@nestjs/common';
+import { Controller, Body, Param, Res } from '@nestjs/common';
 import { ParseUUIDPipe } from '@nestjs/common';
 import { Response } from 'express';
+import { CreateTrackDto } from './dto/tracks.dto';
 
 @Controller('track')
 export default class TracksController {
@@ -15,6 +16,10 @@ export default class TracksController {
     id: string,
     @Res() res: Response,
   ) {
-    this.TracksService.findById(id, res);
+    return this.TracksService.findById(id, res);
+  }
+
+  createTrack(@Body() body: CreateTrackDto, @Res() res: Response) {
+    return this.TracksService.createTrack(body, res);
   }
 }
