+++ src/tracks/tracks.service.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@nestjs/common';
-import { Track } from './dto/tracks.dto';
+import { Track } from './interfaces/track.interface';
 import { Response } from 'express';
 import { randomUUID } from 'crypto';
+import { CreateTrackDto } from './dto/tracks.dto';
 import existById from 'src/helpers/checkExist';
 import ResponseHelper from 'src/helpers/responseHelper';
 import db from 'src/data/inMemoryDB';
@@ -20,5 +21,14 @@ export default class TracksService {
     if (!track) {
       return ResponseHelper.sendNotFound(res, 'Track not found');
     }
+  }
+
+  createTrack(body: CreateTrackDto, res: Response) {
+    const { name, artistId, albumId, duration } = body;
+
+    const data = { id: randomUUID(), name, artistId, albumId, duration };
+
+    this.tracks.push(data);
+    return res.status(201).json(data);
   }
 }
